requirements.txt
aioredis==2.0.1
confluent_kafka==1.8.2
python-dotenv==0.20.0
uvloop==0.16.0
websockets==10.1
